keras-mnist-model/requirements.txt
absl-py==0.7.0
astor==0.7.1
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
gast==0.2.2
grpcio==1.14.2
grpcio-tools==1.14.2
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
numpy==1.16.1
opencv-python==4.0.0.21
protobuf==3.6.1
PyYAML==3.13
scipy==1.2.1
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1

